requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
accelerate==0.23.0
    # via detectors
adversarial-robustness-toolbox==1.16.0
    # via adversarial-insight-ml (pyproject.toml)
aiohttp==3.8.6
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
alembic==1.12.0
    # via optuna
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
arrow==1.3.0
    # via lightning
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
backoff==2.2.1
    # via lightning
beautifulsoup4==4.12.2
    # via lightning
blessed==1.20.0
    # via inquirer
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   lightning
    #   lightning-cloud
    #   uvicorn
cmaes==0.10.0
    # via optuna
colorlog==6.7.0
    # via optuna
contourpy==1.1.1
    # via matplotlib
croniter==1.4.1
    # via lightning
cycler==0.12.1
    # via matplotlib
dateutils==0.6.12
    # via lightning
deepdiff==6.6.0
    # via lightning
detectors==0.1.10
    # via adversarial-insight-ml (pyproject.toml)
exceptiongroup==1.1.3
    # via anyio
faiss-cpu==1.7.4
    # via detectors
fastapi==0.103.2
    # via
    #   lightning
    #   lightning-cloud
filelock==3.12.4
    # via
    #   huggingface-hub
    #   torch
    #   triton
fonttools==4.43.1
    # via matplotlib
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.9.2
    # via
    #   huggingface-hub
    #   lightning
    #   pytorch-lightning
    #   torch
greenlet==3.0.0
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
huggingface-hub==0.17.3
    # via
    #   accelerate
    #   timm
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
imageio==2.31.5
    # via scikit-image
importlib-resources==6.1.0
    # via matplotlib
inquirer==3.1.3
    # via lightning
itsdangerous==2.1.2
    # via starsessions
jinja2==3.1.2
    # via
    #   lightning
    #   torch
joblib==1.3.2
    # via scikit-learn
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.3
    # via scikit-image
lightning==2.0.9.post0
    # via adversarial-insight-ml (pyproject.toml)
lightning-cloud==0.5.39
    # via lightning
lightning-utilities==0.9.0
    # via
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
littleutils==0.2.2
    # via outdated
mako==1.2.4
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
matplotlib==3.8.0
    # via detectors
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
networkx==3.1
    # via
    #   scikit-image
    #   torch
numpy==1.26.0
    # via
    #   accelerate
    #   adversarial-robustness-toolbox
    #   cmaes
    #   contourpy
    #   detectors
    #   imageio
    #   lightning
    #   matplotlib
    #   ogb
    #   optuna
    #   pandas
    #   pytorch-lightning
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
    #   torchmetrics
    #   torchvision
    #   wilds
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.18.1
    # via torch
nvidia-nvjitlink-cu12==12.2.140
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
ogb==1.3.6
    # via wilds
optuna==3.3.0
    # via detectors
ordered-set==4.1.0
    # via deepdiff
outdated==0.2.2
    # via
    #   ogb
    #   wilds
packaging==23.2
    # via
    #   accelerate
    #   huggingface-hub
    #   lightning
    #   lightning-utilities
    #   matplotlib
    #   optuna
    #   pytorch-lightning
    #   scikit-image
pandas==2.1.1
    # via
    #   detectors
    #   ogb
    #   wilds
pillow==10.0.1
    # via
    #   detectors
    #   imageio
    #   matplotlib
    #   scikit-image
    #   torchvision
    #   wilds
psutil==5.9.5
    # via
    #   accelerate
    #   detectors
    #   lightning
pydantic==2.1.1
    # via
    #   fastapi
    #   lightning
pydantic-core==2.4.0
    # via pydantic
pygments==2.16.1
    # via rich
pyjwt==2.8.0
    # via lightning-cloud
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   arrow
    #   croniter
    #   dateutils
    #   matplotlib
    #   pandas
python-editor==1.0.4
    # via inquirer
python-multipart==0.0.6
    # via
    #   lightning
    #   lightning-cloud
pytorch-lightning==2.0.9.post0
    # via lightning
pytz==2023.3.post1
    # via
    #   dateutils
    #   pandas
    #   wilds
pyyaml==6.0.1
    # via
    #   accelerate
    #   huggingface-hub
    #   lightning
    #   optuna
    #   pytorch-lightning
    #   timm
readchar==4.0.5
    # via inquirer
requests==2.31.0
    # via
    #   fsspec
    #   huggingface-hub
    #   lightning
    #   lightning-cloud
    #   outdated
    #   torchvision
rich==13.6.0
    # via
    #   lightning
    #   lightning-cloud
safetensors==0.4.0
    # via timm
scikit-image==0.22.0
    # via detectors
scikit-learn==1.1.3
    # via
    #   adversarial-robustness-toolbox
    #   detectors
    #   ogb
    #   wilds
scipy==1.11.3
    # via
    #   adversarial-robustness-toolbox
    #   detectors
    #   scikit-image
    #   scikit-learn
    #   wilds
six==1.16.0
    # via
    #   adversarial-robustness-toolbox
    #   blessed
    #   lightning-cloud
    #   ogb
    #   python-dateutil
sniffio==1.3.0
    # via anyio
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.21
    # via
    #   alembic
    #   optuna
starlette==0.27.0
    # via
    #   fastapi
    #   lightning
    #   starsessions
starsessions==1.3.0
    # via lightning
sympy==1.12
    # via torch
threadpoolctl==3.2.0
    # via scikit-learn
tifffile==2023.9.26
    # via scikit-image
timm==0.8.19.dev0
    # via detectors
torch==2.1.0
    # via
    #   accelerate
    #   adversarial-insight-ml (pyproject.toml)
    #   detectors
    #   lightning
    #   ogb
    #   pytorch-lightning
    #   timm
    #   torchmetrics
    #   torchvision
    #   wilds
torchmetrics==1.2.0
    # via
    #   adversarial-insight-ml (pyproject.toml)
    #   lightning
    #   pytorch-lightning
torchvision==0.16.0
    # via
    #   detectors
    #   timm
    #   wilds
tqdm==4.66.1
    # via
    #   adversarial-robustness-toolbox
    #   detectors
    #   huggingface-hub
    #   lightning
    #   ogb
    #   optuna
    #   pytorch-lightning
    #   wilds
traitlets==5.11.2
    # via lightning
triton==2.1.0
    # via torch
types-python-dateutil==2.8.19.14
    # via arrow
typing-extensions==4.8.0
    # via
    #   alembic
    #   fastapi
    #   huggingface-hub
    #   lightning
    #   lightning-utilities
    #   pydantic
    #   pydantic-core
    #   pytorch-lightning
    #   sqlalchemy
    #   starlette
    #   torch
    #   uvicorn
tzdata==2023.3
    # via pandas
urllib3==2.0.6
    # via
    #   lightning
    #   lightning-cloud
    #   ogb
    #   requests
uvicorn==0.23.2
    # via
    #   lightning
    #   lightning-cloud
wcwidth==0.2.8
    # via blessed
websocket-client==1.6.3
    # via
    #   lightning
    #   lightning-cloud
websockets==11.0.3
    # via lightning
wilds==2.0.0
    # via detectors
yarl==1.9.2
    # via aiohttp
zipp==3.17.0
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
